# src/main/java/com/whatsapp/Whatsapp_clone/message/MessageRequest.java
package com.whatsapp.Whatsapp_clone.message;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class MessageRequest {

    private String content;
    private String senderId;
    private String receiverId;
    private MessageType type;
    private String chatId;
}